requirements.txt
streamlit
google-generativeai
PyMuPDF
python-dotenv==1.0.1


